requirements.txt
scrapy==2.11.1
python-decouple==3.8
psycopg2==2.9.9
pymongo==4.15.5
itemadapter==0.13.0